requirements.txt
numpy
pandas
matplotlib
scipy
plotly
seaborn
jupyter
notebook
librosa
beautifulsoup4
selenium
python-dotenv
scikit-learn
tensorflow
pydot
visualkeras
opencv-python
opencv-python-headless
soundfile